Move date functions out of 'string.c' and into the new 'date_str.c'.
