+++ lib/Makefile
@@ -12,13 +12,13 @@ SRCROOT	= ..
 
 HDR 	= dao.h dao.p
 
-SRC	= acl.c       chrono32.c  file.c    isnot.c    radix32.c  shm.c  \
-	  archiv32.c  dl_lib.c    record.c  splay.c    xsort.c           \
-	  attr_lib.c  dns.c       header.c  rfc2047.c  string.c  xwrite.c
+SRC	= acl.c       chrono32.c  file.c    isnot.c    radix32.c   shm.c   \
+	  archiv32.c  dl_lib.c    record.c  splay.c    date_str.c  xsort.c \
+	  attr_lib.c  dns.c       header.c  rfc2047.c  string.c    xwrite.c
 
-OBJ	= acl.o       chrono32.o  file.o    isnot.o    radix32.o  shm.o  \
-	  archiv32.o  dl_lib.o    record.o  splay.o    xsort.o           \
-	  attr_lib.o  dns.o       header.o  rfc2047.o  string.o  xwrite.o
+OBJ	= acl.o       chrono32.o  file.o    isnot.o    radix32.o   shm.o   \
+	  archiv32.o  dl_lib.o    record.o  splay.o    date_str.o  xsort.o \
+	  attr_lib.o  dns.o       header.o  rfc2047.o  string.o    xwrite.o
 
 .c.o:   ;$(CC) $(CFLAGS) -c $*.c
